Remove #[allow(dead_code)] annotations and the dead code they masked
checksum, header, and tiles were already fully used and didn't need
the allow. The remaining dead code was genuinely unused outside of
tests, so it's removed:

- TreeNode::internal/tree_to_lut in huffman.rs, superseded by the
  flat-arena compute_codes implementation
- get_best_level_full in fpl/mod.rs, a fallback never wired into
  compress_byte_plane
- encode_rle/encode_fpl_huffman in fpl/compression.rs and
  encode_huffman/encode_huffman_with_codec in huffman.rs, duplicate
  wrappers around logic compress_buffer and
  encode_fpl_huffman_with_histo_bounded/encode_huffman_with_codec_histo
  already implement directly

Their unit tests now exercise the real production functions instead,
so round-trip coverage for RLE/PackBits/Huffman is preserved.

Also applies cargo fmt and fixes the clippy warnings surfaced along
the way: uses the MAX_RATIO constant instead of duplicating the 1.02
literal at each check_ratio call site in compression_ratio.rs, and
drops redundant u32 -> u32 casts in cpp_cross_validation.rs.

cargo fmt --check and cargo clippy --all-features --all-targets
-- -D warnings are both clean.

+++ tests/cpp_cross_validation.rs
@@ -1073,8 +1073,8 @@ fn cpp_encode_rust_decode_f32_lossless_fpl() {
     );
 
     let image = lerc::decode(&blob).unwrap();
-    assert_eq!(image.width, width as u32);
-    assert_eq!(image.height, height as u32);
+    assert_eq!(image.width, width);
+    assert_eq!(image.height, height);
     assert_eq!(image.data_type, DataType::Float);
 
     let decoded = image.as_typed::<f32>().unwrap();
